requirements.txt
pillow==10.2.0
requests==2.31.0
uvicorn==0.27.0
